requirements.txt
# Core dependencies
python-dotenv==1.0.0

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Geospatial
earthengine-api==0.1.384
geopandas==0.14.1
rasterio==1.3.9
shapely==2.0.2
pyproj==3.6.1
fiona==1.9.5

# Data processing
numpy==1.26.2
pandas==2.1.3
xarray==2023.11.0
netCDF4==1.6.5
h5py==3.10.0

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
torch==2.1.1
pytorch-forecasting==1.0.0
pytorch-lightning==2.1.2

# Database
oracledb==2.0.0
sqlalchemy==2.0.23
geoalchemy2==0.14.2

# Orchestration
apache-airflow==2.7.3
apache-airflow-providers-oracle==3.8.0

# Reporting
reportlab==4.0.7
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Utilities
requests==2.31.0
tqdm==4.66.1
python-multipart==0.0.6
aiofiles==23.2.1

# Cloud (OCI)
oci==2.119.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0
